Use experimental API to write remote files.
This is built on top of https://github.com/Microsoft/vscode/pull/30337.

Note that I had to follow the instructions at
https://github.com/Microsoft/vscode/issues/13394 to be able to run
this extension from the OSS version of VS Code.
Unfortunately, that means I made a personal modification to
`.vscode/launch.json`, so hopefully I can find some way to avoid that.

This is pretty exciting in that I had to make minimal changes to my
existing extension to get this to work! The only issue that I'm running
into right now is that `resolveContents()` appears to be called more
often than I expect. It is called twice when I open the file for the
first time. I also opened a remote `.md` file using this mechanism. I was
able to use `Markdown: Open Preview` on this remote file, but when I
closed the preview, `resolveContents()` was called again and I'm not sure
why. Because loading remote file contents could be expensive, we would
ideally fix things so this is called sparingly, or we'll have to maintain
a local cache to avoid extra roundtrips.

Note that I had to launch the OSS version (built from source) using
`./scripts/code.sh --remote nuclide`. There is a TODO in the PR to
eliminate the need to pass this flag.